Fixed building
Added all the necessary methods to each of the classes. If it's still not working, shoot me another message and i shall just incorporate it all into main game. Didnt want to do that right away because then it tells me we have duplicate classes and i still want them separate so its eaiser to look at and deal with